Don't set CPU quota if it's at 0
Some kernels, like raspbian, don't support cpu_cfs cgroups. If the
quota is set to 0, just don't set the cgroup at all rather than
setting it to -1. This has the side effect of not resetting the
cgroup if it was inherited from a parent cgroup, but oh well;
quick and dirty does the job >:)